--- src/defines.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#define EMIN 1e-5
-#define KMAX 30
-
-// Rodar caso teste 5x5
-// #define TESTCASE
-
-
