--- src/grasp.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "common.h"
-
-Solution Construct();
